Move get_attributed_metrics_by_domain
